better modularization
